memento-db: extract mementoDb-addConvXchg.ts from mementoDb-mems.ts
